force ipv4 for prebuild request due to simple-get request timeout when encountering broken ipv6 routers
The underlying issue is that when dns resolves to multiple IPs all ips are attempted. ipv6 is tried first, then ipv4. for every ip failure, it will emit a timeout event which simple-get interprets as an error. This may be a bug in either node (since the entire request hasnt timed out, just 1 ip attempt), or a bug in simple-get where it should be more discerning about which request timeout events it should observe. In any case, this is an issue with a end user machines ipv6 network configuration. Requiring binaries be hosted on a ipv4 network seems reasonable.